extern/nd_sd/requirements.txt
absl-py==2.0.0
accelerate==0.24.1
addict==2.4.0
aiofiles==23.2.1
aiohttp==3.8.6
aiosignal==1.3.1
albumentations==0.4.3
aliyun-python-sdk-core==2.14.0
aliyun-python-sdk-kms==2.16.2
altair==5.1.2
annotated-types==0.6.0
ansi2html==1.8.0
antlr4-python3-runtime==4.8
anyio==3.7.1
asttokens==2.4.0
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
blinker==1.6.3
braceexpand==0.1.7
cachetools==5.3.1
carvekit-colab==4.1.0
certifi==2023.7.22
cffi==1.16.0
cfgv==3.4.0
charset-normalizer==3.3.0
click==8.1.7
cmake==3.27.6
comm==0.1.4
ConfigArgParse==1.7
contourpy==1.1.1
crcmod==1.7
cryptography==41.0.5
cycler==0.12.1
dash==2.14.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dataclasses-json==0.6.1
datasets==2.4.0
decorator==5.1.1
Deprecated==1.2.14
diffusers==0.17.0
dill==0.3.5.1
distlib==0.3.7
easydict==1.11
einops==0.3.0
exceptiongroup==1.1.3
executing==2.0.0
expecttest==0.1.6
fairscale==0.4.13
fastapi==0.103.2
fastcore==1.5.29
fastjsonschema==2.18.1
ffmpy==0.3.1
filelock==3.12.4
fire==0.4.0
flake8==6.1.0
flash-attn==1.0.4
Flask==3.0.0
fonttools==4.43.1
freetype-py==2.4.0
frozenlist==1.4.0
fsspec==2023.9.2
ftfy==6.1.1
future==0.18.3
gitdb==4.0.10
GitPython==3.1.37
google-auth==2.23.3
google-auth-oauthlib==1.0.0
gradio==3.13.2
grpcio==1.59.0
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
huggingface-hub==0.18.0
identify==2.5.33
idna==3.4
imageio==2.9.0
imageio-ffmpeg==0.4.2
imgaug==0.2.6
importlib-metadata==6.8.0
importlib-resources==6.1.0
invisible-watermark==0.2.0
ipython==8.16.1
ipywidgets==8.1.1
isort==5.12.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
jmespath==0.10.0
joblib==1.3.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter_core==5.5.0
jupyterlab-widgets==3.0.9
kiwisolver==1.4.5
kornia==0.6.0
lazy_loader==0.3
lightning-utilities==0.9.0
linkify-it-py==2.0.2
lit==17.0.2
loguru==0.7.2
lovely-numpy==0.2.9
lovely-tensors==0.1.15
Markdown==3.5
markdown-it-py==2.2.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.8.0
matplotlib-inline==0.1.6
mccabe==0.7.0
mdit-py-plugins==0.3.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.13
mypy-extensions==1.0.0
nbformat==5.7.0
nest-asyncio==1.5.8
networkx==3.1
nodeenv==1.8.0
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
omegaconf==2.1.1
open-clip-torch==2.20.0
open3d==0.17.0
opencv-python==4.5.5.64
orjson==3.9.9
oss2==2.18.3
packaging==23.2
pandas==2.1.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.1
platformdirs==3.11.0
plotly==5.13.1
pre-commit==3.5.0
prompt-toolkit==3.0.39
protobuf==3.20.3
psutil==5.9.6
ptyprocess==0.7.0
pudb==2019.2
pure-eval==0.2.2
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.11.1
pycparser==2.21
pycryptodome==3.19.0
pydantic==2.4.2
pydantic_core==2.10.1
pydeck==0.8.1b0
pyDeprecate==0.3.1
pydub==0.25.1
pyflakes==3.1.0
pyglet==2.0.9
pygltflib==1.16.1
Pygments==2.16.1
PyOpenGL==3.1.0
pyparsing==3.1.1
pyquaternion==0.9.9
pyrender==0.1.45
python-dateutil==2.8.2
python-multipart==0.0.6
pytorch-lightning==1.9.0
pytz==2023.3.post1
PyWavelets==1.4.1
PyYAML==6.0.1
referencing==0.30.2
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
responses==0.18.0
retrying==1.3.4
rich==13.6.0
rpds-py==0.10.6
rsa==4.9
safetensors==0.4.0
scikit-image==0.20.0
scikit-learn==1.3.2
scipy==1.9.1
sentencepiece==0.1.99
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
stack-data==0.6.3
starlette==0.27.0
streamlit==1.27.2
sympy==1.12
taming-transformers-rom1504==0.0.6
tenacity==8.2.3
tensorboard==2.14.1
tensorboard-data-server==0.7.1
termcolor==2.3.0
test-tube==0.7.5
threadpoolctl==3.2.0
tifffile==2023.9.26
timm==0.6.13
tokenizers==0.12.1
toml==0.10.2
toolz==0.12.0
torch==2.1.0+cu118
torch-fidelity==0.3.0
torchmetrics==1.2.0
torchsde==0.2.6
torchvision==0.16.0+cu118
tornado==6.3.3
tqdm==4.66.1
traitlets==5.11.2
trampoline==0.1.2
transformations==2022.9.26
transformers==4.27.1
trimesh==4.0.2
triton==2.1.0
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.1
uc-micro-py==1.0.2
urllib3==2.0.6
urwid==2.2.2
uvicorn==0.23.2
validators==0.22.0
virtualenv==20.25.0
watchdog==3.0.0
wcwidth==0.2.8
webdataset==0.2.5
websockets==11.0.3
Werkzeug==3.0.0
widgetsnbextension==4.0.9
wrapt==1.15.0
xxhash==3.4.1
yapf==0.30.0
yarl==1.9.2
zipp==3.17.0
modelscope==1.9.5